update disable mouseHover an touch devices
# Conflicts:
#	modules/mouseHover/view.js

+++ modules/mouseHover/view.js
@@ -1,14 +1,40 @@
 import MouseHoverPopup from "./model";
 
-const MouseHoverPopupView = Backbone.View.extend({
-    initialize: function (attr) {
-        this.model = new MouseHoverPopup(attr);
-        this.listenTo(this.model, {
-            "render": this.render,
-            "destroy": this.destroy
-        });
+const MouseHoverPopupView = Backbone.View.extend(/** @lends MouseHoverPopupView.prototype */{
+    /**
+     * @class MouseHoverPopupView
+     * @extends Backbone.View
+     * @memberof MouseHover
+     * @constructs
+     * @classdesc Shows an ol.overlay with configured informations about the hovered feature on the map. It's always disabled on touch devices because they usually don't have a mouse to trigger a hover event.
+     * @fires Util#RadioRequestUtilIsAny
+     * @listens MouseHover#render
+     * @listens MouseHover#destroy
+     */
+    initialize: function () {
+        if (!this.isTouchdevice()) {
+            this.model = new MouseHoverPopup();
+            this.listenTo(this.model, {
+                "render": this.render,
+                "destroy": this.destroy
+            });
 
-        this.createOverlayElement();
+            this.createOverlayElement();
+        }
+    },
+
+    /**
+     * Checks the device type. Disable the ability on smartphones and tablets as they usually don't have a mouse to hover.
+     * @fires Util#RadioRequestUtilIsAny
+     * @returns {Boolean}
+     */
+    isTouchdevice: function () {
+        if (Radio.request("Util", "isAny") === null) {
+            return false;
+        }
+        else {
+            return true;
+        }
     },
 
     /**
